+++ python/GafferUI/EditScopeUI.py
@@ -150,9 +150,34 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 		# run the default dropSignal handler from PlugValueWidget.
 		self.dropSignal().connectFront( Gaffer.WeakMethod( self.__drop ), scoped = False )
 
+		self.__updatePlugInputChangedConnection()
+		self.__acquireContextTracker()
+
 	def hasLabel( self ) :
 
 		return True
+
+	def setPlugs( self, plugs ) :
+
+		GafferUI.PlugValueWidget.setPlugs( self, plugs )
+		self.__updatePlugInputChangedConnection()
+		self.__acquireContextTracker()
+
+	def getToolTip( self ) :
+
+		editScope = self.__editScope()
+		if editScope is None :
+			return "Edits will be made using the last relevant node, including nodes not in any EditScope."
+
+		name = editScope.relativeName( editScope.scriptNode() )
+		inputNode = self.__inputNode()
+		if inputNode is None :
+			return f"{name} cannot be used while nothing is viewed."
+		elif editScope not in self.__activeEditScopes() :
+			inputNodeName = inputNode.relativeName( inputNode.scriptNode() )
+			return f"{name} cannot be used as it is not upstream of {inputNodeName}."
+		else :
+			return f"Edits will be made in {name}."
 
 	# We don't actually display values, but this is also called whenever the
 	# input changes, which is when we need to update.
@@ -160,7 +185,7 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 
 		editScope = self.__editScope()
 		editScopeActive = editScope is not None
-		self.__updateMenuButton( editScope )
+		self.__updateMenuButton()
 		self.__navigationMenuButton.setEnabled( editScopeActive )
 		if editScopeActive :
 			self.__editScopeNameChangedConnection = editScope.nameChangedSignal().connect(
@@ -177,19 +202,50 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 			self._qtWidget().setProperty( "editScopeActive", GafferUI._Variant.toVariant( editScopeActive ) )
 			self._repolish()
 
-	def __updateMenuButton( self, editScope ) :
+	def __updatePlugInputChangedConnection( self ) :
 
+		self.__plugInputChangedConnection = self.getPlug().node().plugInputChangedSignal().connect(
+			Gaffer.WeakMethod( self.__plugInputChanged ), scoped = True
+		)
+
+	def __plugInputChanged( self, plug ) :
+
+		if plug.getName() == "in" and plug.parent() == self.getPlug().node() :
+			# The result of `__inputNode()` will have changed.
+			self.__acquireContextTracker()
+			self.__updateMenuButton()
+
+	def __acquireContextTracker( self ) :
+
+		self.__contextTracker = GafferUI.ContextTracker.acquire( self.__inputNode() )
+		self.__contextTrackerChangedConnection = self.__contextTracker.changedSignal().connect(
+			Gaffer.WeakMethod( self.__contextTrackerChanged ), scoped = True
+		)
+
+	def __updateMenuButton( self ) :
+
+		editScope = self.__editScope()
 		self.__menuButton.setText( editScope.getName() if editScope is not None else "None" )
-		self.__menuButton.setImage( self.__editScopeSwatch( editScope ) if editScope is not None else None )
+
+		if editScope is not None :
+			self.__menuButton.setImage(
+				self.__editScopeSwatch( editScope ) if self.__contextTracker.isActive( editScope ) else "warningSmall.png"
+			)
+		else :
+			self.__menuButton.setImage( None )
 
 	def __editScopeNameChanged( self, editScope, oldName ) :
 
-		self.__updateMenuButton( editScope )
+		self.__updateMenuButton()
 
 	def __editScopeMetadataChanged( self, editScope, key, reason ) :
 
 		if key == "nodeGadget:color" :
-			self.__updateMenuButton( editScope )
+			self.__updateMenuButton()
+
+	def __contextTrackerChanged( self, contextTracker ) :
+
+		self.__updateMenuButton()
 
 	def __editScope( self ) :
 
@@ -230,6 +286,20 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 				return output.node()
 
 		return inputNode
+
+	def __activeEditScopes( self ) :
+
+		node = self.__inputNode()
+		if node is None :
+			return []
+
+		result = Gaffer.NodeAlgo.findAllUpstream( node, self.__editScopePredicate )
+		if self.__editScopePredicate( node ) :
+			result.insert( 0, node )
+
+		result = [ n for n in result if self.__contextTracker.isActive( n ) ]
+
+		return result
 
 	def __buildMenu( self, result, path, currentEditScope ) :
 
@@ -290,16 +360,10 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 		if self.getPlug().getInput() is not None :
 			currentEditScope = self.getPlug().getInput().parent()
 
-		node = self.__inputNode()
-		if node is not None :
-			upstream = Gaffer.NodeAlgo.findAllUpstream( node, self.__editScopePredicate )
-			if self.__editScopePredicate( node ) :
-				upstream.insert( 0, node )
+		activeEditScopes = self.__activeEditScopes()
 
-			downstream = Gaffer.NodeAlgo.findAllDownstream( node, self.__editScopePredicate )
-		else :
-			upstream = []
-			downstream = []
+		node = self.__inputNode()
+		downstream = Gaffer.NodeAlgo.findAllDownstream( node, self.__editScopePredicate ) if node is not None else []
 
 		# Each child of the root will get its own section in the menu
 		# if it has children. The section will be preceded by a divider
@@ -321,13 +385,11 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 				currentNode = currentNode.setdefault( n.getName(), {} )
 			currentNode[editScope.getName()] = editScope
 
-		if upstream :
-			for editScope in sorted( upstream, key = lambda e : e.relativeName( e.scriptNode() ) ) :
-				addToMenuHierarchy( editScope, "Upstream" )
+		for editScope in reversed( activeEditScopes ) :
+			addToMenuHierarchy( editScope, "Upstream" )
 
-		if downstream :
-			for editScope in sorted( downstream, key = lambda e : e.relativeName( e.scriptNode() ) ) :
-				addToMenuHierarchy( editScope, "Downstream" )
+		for editScope in sorted( downstream, key = lambda e : e.relativeName( e.scriptNode() ) ) :
+			addToMenuHierarchy( editScope, "Downstream" )
 
 		menuPath = Gaffer.DictPath( menuHierarchy, "/" )
 
@@ -456,11 +518,7 @@ class EditScopePlugValueWidget( GafferUI.PlugValueWidget ) :
 					GafferUI.Label( "<h4>The Edit Scope cannot be set while nothing is viewed</h4>" )
 			self.__popup.popup( parent = self )
 		elif dropNode :
-			upstream = Gaffer.NodeAlgo.findAllUpstream( inputNode, self.__editScopePredicate )
-			if self.__editScopePredicate( inputNode ) :
-				upstream.insert( 0, inputNode )
-
-			if dropNode in upstream :
+			if dropNode in self.__activeEditScopes() :
 				self.__connectEditScope( dropNode )
 			else :
 				with GafferUI.PopupWindow() as self.__popup :